# backend/src/swagger/openapi.yaml
openapi: '3.0.0'
servers: 
  - url: https://gipsa-solutionchallenge2022.du.r.appspot.com
  - url: http://34.64.244.115:8080
info:
  description: This is a simple API
  version: "1.0.0"
  title: G.Ipsa API
  license:
    name: Apache 2.0
...
paths:
...
  /follow:
    get:
      tags: 
        - follow
      summary: followerId 넣으면 본인이 팔로하는 사람, followedUserId라면 본인을 팔로하는 사람
      parameters: 
      - in: query
        name: followerId
        schema:
          type: number
      - in: query
        name: creatorId
        schema:
          type: number
      responses:
        '200':
          description: return
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Follow'
  /follow/click:
    post:
      tags: 
        - follow
      summary: 팔로우 되어있으면 팔로 아니면 언팔
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Follow'
      responses:
        default:
          description: good
  /comment:
    get:
      tags:
        - comment
      summary: bring comments
      parameters: 
      - in: query
        name: userId
        schema:
          type: number
      - in: query
        name: storyId
        schema:
          type: number
      responses:
        '200':
          description: array of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '400':
          description: invalid
  /comment/create:
    post:
      tags: 
        - comment
      summary: 댓글 추가
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        default:
          description: successful operation 
...
components:
  schemas:
...
    Comment:
      type: object
      required:
        - userId
        - storyId
        - content
      properties:
        id:
          type: number
          format: int64
        userId:
          type: number
          format: int64
        storyId:
          type: number
          format: int64
        content:
          type: string
        createdAt:
          type: string
          format: date-time
          example: '2016-08-29T09:12:33.001Z'
    Follow:
      type: object
      required:
        - followerId
        - creatorId
      properties:
        followerId:
          type: number
          format: int64
        creatorId:
          type: number
          format: int64
...